move continuation support into its own header
